Switching game style
The player and their weapon will always be on the left matrix. Enemies will spawn from the rightmost column of the second matrix and move towards the player. The enemies can either move forward, forward and up, or forward and down. I might make it so that they only move forward, since it seems a little too hard right now.